Moved configuration decorators to `configs.py` module.
