+++ chalice_restful/__init__.py
@@ -1,2 +1,3 @@
-from .core import Api, Resource, route, config, flag
+from .core import Api, Resource, route
+from .configs import config, flag
 from .authorization import authorizer
